requirements.txt
# Core Dependencies
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==1.0.2
langchain-text-splitters==0.0.2

# OpenAI
openai==2.7.2

# Vector Database
chromadb==1.3.4

# PDF Processing
pypdf==6.2.0
pypdfium2==4.30.0

# Environment Variables
python-dotenv==1.0.0

# API Server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Optional: For enhanced functionality
# langchain-experimental==0.4.0  # Uncomment if needed
